+++ packages/sdk/src/utils/reservoirChains.ts
@@ -490,6 +490,14 @@ const game7 = {
   checkPollingInterval: 1000,
 } satisfies ReservoirBaseChain
 
+const creatorTestnet = {
+  id: 4654,
+  name: 'Creator Testnet',
+  baseApiUrl: 'https://api-creator-testnet.reservoir.tools',
+  paymentTokens: chainPaymentTokensMap[4654],
+  checkPollingInterval: 1000,
+} satisfies ReservoirBaseChain
+
 export const reservoirChains = {
   mainnet,
   polygon,
@@ -551,4 +559,5 @@ export const reservoirChains = {
   animeTestnet,
   monadDevnet,
   game7,
+  creatorTestnet,
 }